refactor: move changePassInput component to components folder
