+++ src/components/mainLayout/changePassInput/changePassInput.tsx
@@ -1,23 +1,24 @@
 import { useState } from "react";
 import { useController } from "react-hook-form";
-
 import cn from "classnames";
-
 import Image from "next/image";
 
-import { IChangePassInput, TChangePassForm } from "../../types/common/UiKitProps";
-import { InputTypeList } from "../../helpers/Input";
-import showPassword from "../../assets/pages/signUp/showPassword.svg";
+import { IChangePassInput, TChangePassForm } from "../../../types/common/UiKitProps";
+import { InputTypeList } from "../../../helpers/Input";
+
+import showPassword from "../../../assets/pages/signUp/showPassword.svg";
 
 import styles from "./changePassInput.module.scss";
 
 const ChangePassInput = ({ label, type, placeholder, autoComplete, subtitle, error, ...props }: IChangePassInput) => {
 	const { field, fieldState } = useController<TChangePassForm>(props);
 	const [passwordType, setPasswordType] = useState<InputTypeList>(InputTypeList.Password);
+
 	const togglePasswordVisibility = () =>
 		setPasswordType(passwordType === InputTypeList.Password ? InputTypeList.Text : InputTypeList.Password);
 
-	const value = typeof field.value === "boolean" ? String(field.value) : field.value;
+	const inputValue = typeof field.value === "boolean" ? String(field.value) : field.value;
+
 	return (
 		<div className={styles.inputWrap}>
 			<label className={styles.inputWrap__label}>{label}</label>
@@ -28,7 +29,7 @@ const ChangePassInput = ({ label, type, placeholder, autoComplete, subtitle, err
 					placeholder={placeholder}
 					className={styles.inputWrap__input}
 					autoComplete={autoComplete}
-					value={value}
+					value={inputValue}
 				/>
 				{type === InputTypeList.Password && (
 					<button onClick={togglePasswordVisibility} className={styles.inputWrap__passwordEye} type="button">
